+++ src/api/controllers.ts
@@ -1,6 +1,6 @@
 import type { Got } from 'got' with { 'resolution-mode': 'require' }
 import type { Device } from '../types/device/Device'
-import type { Controller } from '../types/device/Controller'
+import type { Controller, GenericSwitch } from '../types/device/Controller'
 
 export default (got: Got) => {
   return {
@@ -15,6 +15,26 @@ export default (got: Got) => {
         throw new Error('The requested device is not a controller')
       }
       return device as Controller
+    },
+
+    async setControlMode({
+      id,
+      controlMode,
+    }: {
+      id: string
+      controlMode: GenericSwitch['attributes']['controlMode']
+    }) {
+      await got
+        .patch(`devices/${id}`, {
+          json: [
+            {
+              attributes: {
+                controlMode,
+              },
+            },
+          ],
+        })
+        .json()
     },
   }
 }