+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased](https://github.com/lpgera/dirigera/compare/v1.7.0...HEAD)
 
 - Add BILRESA control mode changing endpoint.
+- Fix Hub types.
+-
 
 ## [1.7.0](https://github.com/lpgera/dirigera/compare/v1.6.0...v1.7.0) - 2025-12-27